Parse basic RSS XML in doctest
For info: it's a good idea to place the `~S` sigil at the start
of the docstring e.g.
```
@doc ~S"""
...
"""
```

It stops `\r\n` characters from being converted to carriage returns and
newlines before the test is started. Without the sigil there was an
error where the variable `xml` was being treated as a function that
could not be found.

https://elixir-lang.org/getting-started/sigils.html